requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain & LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-aws>=0.1.0
langchain-mongodb>=0.1.0

# MongoDB
pymongo>=4.6.0
certifi>=2024.0.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
aiohttp>=3.9.0
